chore: Fix pre-commit hook style autofixing for client
Prettier wasn't being executed on the client code because of
create-react-app overwriting the global eslint config.
In the future ejecting might be an option, but for now I opted to ignore
the client directory in the global eslint config and run prettier
directly on everything in the client folder.
This commit includes the changes to the npm scripts, as well as
all style changes prettier made.